quiz-generator: feature to update quiz database and lastUsed field
Add feature to choose if the quiz generatior will affect the
database. If this feature is set on true, lastUsed field of
every question in the genertated quiz will be updated to current
date. Also, having this feature on will create a new document in
the quiz database containing all the IDs of the questions generated
in the quiz.

Signed-off-by: Costin-Robert Sin <sin.costinrobert@gmail.com>